@@ -4,26 +4,26 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
-    "@supabase/supabase-js": "^1.33.2",
+    "@supabase/supabase-js": "^1.35.3",
     "@types/color": "^3.0.3",
-    "axios": "^0.26.1",
-    "color": "^4.2.1",
+    "axios": "^0.27.2",
+    "color": "^4.2.3",
     "command-line-args": "^5.2.1",
-    "dayjs": "^1.11.0",
+    "dayjs": "^1.11.2",
     "faunadb": "^4.5.4",
     "lodash": "^4.17.21",
     "rxjs": "^7.5.5",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.6.3",
-    "winston": "^3.6.0"
+    "ts-node": "^10.8.0",
+    "typescript": "^4.7.2",
+    "winston": "^3.7.2"
   },
   "devDependencies": {
-    "@jest/types": "^27.5.1",
-    "@types/jest": "^27.4.1",
-    "@types/lodash": "^4.14.181",
-    "@types/node": "^17.0.23",
-    "jest": "^27.5.1",
-    "ts-jest": "^27.1.4",
+    "@jest/types": "^28.1.0",
+    "@types/jest": "^27.5.1",
+    "@types/lodash": "^4.14.182",
+    "@types/node": "^17.0.35",
+    "jest": "^28.1.0",
+    "ts-jest": "^28.0.3",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
     "tslint-config-standard": "^9.0.0"
