@@ -1,5 +1,5 @@
-import { from } from 'rxjs'
-import { mergeMap } from 'rxjs/operators'
+import { from, lastValueFrom } from 'rxjs'
+import { delay, mergeMap, toArray, retry } from 'rxjs/operators'
 
 import { GetDeputesFromNosDeputesFR } from './WrapperNosDeputesFR'
 import { MapGroupeParlementaire } from './Mapping'
@@ -14,23 +14,65 @@ import {
   CreateGroupeParlementaireToSupabase,
   UpdateGroupeParlementaireToSupabase,
 } from './WrapperSupabase'
+import { GetGroupeParlementaireExplainText } from './WrapperWikipedia'
+
+function delayedResolve<T>(ms, value: T): Promise<T> {
+  return new Promise((resolve) => setTimeout(() => resolve(value), ms))
+}
 
 export async function ManageGroupes() {
+  const groupesFromSupabase: Types.Canonical.GroupeParlementaire[] =
+    await GetGroupesFromSupabase()
+  GetLogger().info('groupesFromSupabase:', groupesFromSupabase)
   const groupesFromNosDeputesFR = await GetDeputesFromNosDeputesFR()
   GetLogger().info('groupesFromNosDeputesFR:', groupesFromNosDeputesFR)
   const canonicalGroupesFromNosDeputesFR = groupesFromNosDeputesFR.map((gp) =>
-    MapGroupeParlementaire(gp)
+    MapGroupeParlementaire(
+      gp,
+      groupesFromSupabase.find((g) => g.Sigle === gp.acronyme)
+    )
   )
   GetLogger().info(
     'canonicalGroupesFromNosDeputesFR:',
     canonicalGroupesFromNosDeputesFR
   )
-  const groupesFromSupabase = await GetGroupesFromSupabase()
-  GetLogger().info('groupesFromFaunaDB:', groupesFromSupabase)
+  const groupeDescriptions = await lastValueFrom(
+    from(groupesFromSupabase).pipe(
+      mergeMap(async (gp) => {
+        if (gp.IDWikipedia === null) {
+          return Promise.resolve({ title: gp.Sigle, desc: '' })
+        }
+
+        return delayedResolve(1000, {
+          title: gp.Sigle,
+          desc: await GetGroupeParlementaireExplainText(
+            encodeURI(gp.IDWikipedia)
+          ),
+        })
+      }, 1),
+      toArray()
+    )
+  )
+  GetLogger().info('groupeDescriptions:', groupeDescriptions)
+  const canonicalGroupesFromNosDeputesFRWithDesc =
+    canonicalGroupesFromNosDeputesFR.map((gp, i) =>
+      Object.assign({}, gp, {
+        DescriptionWikipedia: groupeDescriptions.find(
+          (gpd) => gpd.title === gp.Sigle
+        ).desc,
+      })
+    )
+  GetLogger().info(
+    'canonicalGroupesFromNosDeputesFRWithDesc:',
+    canonicalGroupesFromNosDeputesFRWithDesc
+  )
   const res = CompareLists(
-    canonicalGroupesFromNosDeputesFR,
+    canonicalGroupesFromNosDeputesFRWithDesc,
     groupesFromSupabase,
-    (a, b) => a.Sigle === b.Sigle && a.Actif === b.Actif,
+    (a, b) =>
+      a.Sigle === b.Sigle &&
+      a.Actif === b.Actif &&
+      a.DescriptionWikipedia === b.DescriptionWikipedia,
     'Sigle',
     true
   )
